允许根据 peer 的信息拉黑恶意 BT 用户 (#901)
* create peer filter to allow blocking peer connections

* silly windows

* add `peer_filter_t` and `peer_info_t`

* fix

* fix include

* fix

* update

* jvm peer filter

* fix filter field name

* share ratio limit

* add torrent peer config

* add ip parser module

* fix test

* init peer filter settings page

* add `get_peers` to `torrent_handle_t` and provide more peer info

* peer settings ui

* ip blacklist ui

* fix test

* apply review

* update generated code

* fix test

---------

+++ app/shared/src/commonMain/kotlin/ui/foundation/richtext/RichText.kt
@@ -115,7 +115,7 @@ private fun rememberAnnotatedMaskState(
 @Stable
 object RichTextDefaults {
     val StickerSize: Int = 24
-    val FontSize: Int = 16
+    val FontSize: Float = 16f
 
     @UiThread
     fun checkSanityAndOpen(url: String, context: Context, navigator: BrowserNavigator, toaster: Toaster) {
@@ -227,7 +227,7 @@ object RichTextDefaults {
                         addStyle(
                             style = SpanStyle(
                                 color = textColor,
-                                fontSize = if (e.size.toFloat() != bodyLarge) e.size.sp else 15.5.sp,
+                                fontSize = if (e.size != bodyLarge) e.size.sp else 15.5.sp,
                                 fontWeight = if (e.bold) FontWeight.Bold else null,
                                 fontStyle = if (e.italic) FontStyle.Italic else null,
                                 textDecoration = if (!e.underline && !e.strikethrough) null
